Error reporter

+++ src/main/java/ru/zont/dsbot/core/util/Strings.java
@@ -53,7 +53,7 @@ public class Strings {
         return strLocal != null && strLocal.containsKey(id) || STR_CORE.containsKey(id);
     }
 
-    public String getPlural(int count, String id) {
+    public String getPlural(long count, String id) {
         String one = id + ".one";
         String few = id + ".few";
         String other = id + ".other";
@@ -71,8 +71,8 @@ public class Strings {
         return getPlural(count, get(one), get(few), get(other));
     }
 
-    public static String getPlural(int count, String one, String few, String other) {
-        int c = (count % 100);
+    public static String getPlural(long count, String one, String few, String other) {
+        long c = (count % 100);
         if (c == 1 || (c > 20 && c % 10 == 1))
             return String.format(one, count);
         if ((c < 10 || c > 20) && c % 10 >= 2 && c % 10 <= 4)
